pyproject.toml
[project]
name = "aws-etl-pipeline"
version = "0.1.0"
description = "AWS ETL pipeline project"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "pandas",
    "boto3",
    "sqlalchemy",
    "pymysql",
    "python-dotenv",
    "pyyaml>=6.0.3",
    "ipykernel>=7.2.0",
    "cryptography>=46.0.5",
]
